refactor(storefront/cart): move cart store to shared/lib
