+++ apps/storefront/src/components/layout/store-header.tsx
@@ -7,9 +7,8 @@ import { Icons } from "@dukkani/ui/components/icons";
 import { Skeleton } from "@dukkani/ui/components/skeleton";
 import { usePathname, useRouter } from "next/navigation";
 import { CartDrawer } from "@/components/app/cart-drawer";
-import { useCartHydration } from "@/hooks/use-cart-hydration";
 import { isDetailPage } from "@/shared/config/routes";
-import { useCartStore } from "@/stores/cart.store";
+import { useCartHydration, useCartStore } from "@/shared/lib/cart/store";
 
 interface StoreHeaderProps {
   store: StorePublicOutput;
